NULL -> nullptr

+++ inc/game.hpp
@@ -23,8 +23,8 @@
 namespace tetris {
 
 
-const int WIN_WIDTH = 1500;
-const int WIN_HEIGHT = 1000;
+const size_t WIN_WIDTH = 1500;
+const size_t WIN_HEIGHT = 1000;
 const std::string MY_PATH = "/home/IN204/project";
